5.1.6 Рефакторинг модуля с утилитами. Разобьёт модуль утилит на две части: render.js и common.js. Настроит функции отрисовки render и renderTemplate для работы с компонентами. Добавит функции replace и remove. Вынесет общие функции в модуль common.js
