+++ src/mock/event.js
@@ -1,4 +1,5 @@
 import {EventCategory} from "../const.js";
+import {getRandomInteger, getRandomItem} from "../utils/common.js";
 
 const MIN_SENTENCE_COUNT = 1;
 const MAX_SENTENCE_COUNT = 5;
@@ -10,17 +11,6 @@ const MIN_INTERVAL_BETWEEN_EVENTS = 20;
 const MAX_INTERVAL_BETWEEN_EVENTS = 200;
 const MIN_PHOTO_COUNT = 1;
 const MAX_PHOTO_COUNT = 5;
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomItem = (array) => {
-  return array[getRandomInteger(0, array.length - 1)];
-};
 
 const transferTypes = [
   `Taxi`,
